give all preferences dialogs a clear name
The numbering was somewhat artificial.
