+++ src/Mod/TechDraw/Gui/AppTechDrawGui.cpp
@@ -40,12 +40,12 @@
 #include "MDIViewPage.h"
 
 #include "DlgPrefsTechDrawGeneralImp.h"
-#include "DlgPrefsTechDraw2Imp.h"
+#include "DlgPrefsTechDrawscaleImp.h"
 #include "DlgPrefsTechDrawAnnotationImp.h"
 #include "DlgPrefsTechDrawDimensionsImp.h"
 #include "DlgPrefsTechDrawColorsImp.h"
-#include "DlgPrefsTechDraw4Imp.h"
-#include "DlgPrefsTechDraw5Imp.h"
+#include "DlgPrefsTechDrawAdvancedImp.h"
+#include "DlgPrefsTechDrawHLRImp.h"
 #include "ViewProviderPage.h"
 #include "ViewProviderDrawingView.h"
 #include "ViewProviderDimension.h"
@@ -151,13 +151,13 @@ PyMOD_INIT_FUNC(TechDrawGui)
     TechDrawGui::ViewProviderCosmeticExtension::init();
 
     // register preferences pages
-    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawGeneralImp> ("TechDraw");  //General
-    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDraw2Imp> ("TechDraw");  //Scale
+    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawGeneralImp> ("TechDraw");    //General
+    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawScaleImp> ("TechDraw");      //Scale
     new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawDimensionsImp>("TechDraw");  //Dimensions
     new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawAnnotationImp> ("TechDraw"); //Annotation
-    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawColorsImp>("TechDraw");  //Colors
-    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDraw5Imp> ("TechDraw");  //HLR
-    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDraw4Imp> ("TechDraw");  //Advanced
+    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawColorsImp>("TechDraw");      //Colors
+    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawHLRImp> ("TechDraw");        //HLR
+    new Gui::PrefPageProducer<TechDrawGui::DlgPrefsTechDrawAdvancedImp> ("TechDraw");   //Advanced
 
     // add resources and reloads the translators
     loadTechDrawResource();
